requirements.txt
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pip==21.1
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.24.1
scipy==1.6.3
setuptools==41.2.0
six==1.15.0
sklearn>=0.0
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
wheel==0.36.2
torch>=1.7.1
torchvision>=0.8.2